huge refactor to faciliate animation
svg path changes can only be animated if the before & after paths have
the same number and types of points. so everything has to have the same
number of MX,Y and LX,Y points, with the Ms and Ls in the same positions
in each. in order to make this possible ive rewritten everything to
enable passing in a custom version of the boomshak typeface object, and
written a padTypeface function to do the magic part.

the magic part means calculating the maximum number of strokes of all
glpyhs in the typeface and the maximum number of points in any stroke.
then, all glyphs are changed so that they repeat strokes back and forth
until they match the maximum. the same process is also applied to the
points within each stroke. after this processing is complete, any glyph
using the padded version of the typeface can be animated to any other
glyph.

i needed to get a bit silly here with the metaprogramming to make this
happen and the size of the code has ballooned a bit. didnt really expect
this to be so complex tbh but im glad ive managed to retain the 0
dependencies by implementing custom bits and pieces of lodash-inspired
functionality to support the increased complexity.